+++ src/frontend/apps/impress/src/features/pads/pad-management/api/useRemovePad.tsx
