+++ src/frontend/apps/impress/src/features/pads/pad-management/components/ModalRemovePad.tsx
